feat: Initialize redux store states using the server data
Redux store states for posts, comments and categories were initialized
using hard-coded values. This change eliminates the hard coded values
and uses the data received from the backend to initialize the states.
Posts and categories are initialized by the App and comments state is
lazily updated as the user shows the details of a particular post.